Refactor PositionDetail component and reorganize file structure for position features
