+++ frontend/src/components/PositionDetail.tsx
@@ -1,13 +1,14 @@
 import React from 'react';
 import { useParams } from 'react-router-dom';
+import { KanbanBoard } from '../features/position/components/KanbanBoard';
 
 const PositionDetail = () => {
   const { id } = useParams();
+  const positionId = id ? parseInt(id, 10) : 0;
   
   return (
-    <div className='bg-red-700 p-4 border-2 border-red-900'>
-      <h2 className='text-3xl font-bold underline'>Position Details</h2>
-      <p>Position ID: {id}</p>
+    <div className="min-h-screen bg-gray-100 dark:bg-gray-900">
+      <KanbanBoard positionId={positionId} />
     </div>
   );
 };
